[zshrc] Remove p10k and improve ZSH startup time from 1s to 0.2s
It was ample time to do something about the long startup times. First of
all we got rid of p10k, because we didn't use much of it. We returned to
our own hand-crafted prompt that in itself is much faster.

Furthermore, we now cache command outputs that want to be sourced
immediately like `kubectl completion zsh`.